Refactor: ExternalApiHandler 구현체 패기지 이동
